#326 feat: Get the converter once when converting chart datasets
